+++ src/mc-audio-config.cmake
@@ -3,8 +3,5 @@
 include(CMakeFindDependencyMacro)
 
 find_dependency(mc-core REQUIRED)
-find_dependency(drflac REQUIRED)
-find_dependency(drmp3 REQUIRED)
-find_dependency(drwav REQUIRED)
 
 include("${CMAKE_CURRENT_LIST_DIR}/mc-audio-targets.cmake")